Implement FieldAccess::GetDecl when base != NULL
Added the some code the implement the part of the FieldAccess::GetDecl method
that is executed when a field access's base portion is not NULL. This involved
adding a few getters to the Type classes, VarDecl class, and the ClassDecl
class. Also, implemented the FieldAccess::GetType method in terms of
FieldAccess::GetDecl.