Fixed bug in Expr::GetFieldDecl(Identifier, Expr)
There was a possible infinite loop condition in the
Expr::GetFieldDecl(Identifier, Expr) method. When it is called, the 'this'
pointer is an instance of the Expr class. Thus executing
GetFieldDecl(field, this) presents recursion when this is not the intended
behaviour. Instead, the Expr::GetFieldDecl(Identifier, Node) method is meant to
be called here. Thus static_cast's were introduced to inform the compiler which
function it should call.