Add code generation methods for ArrayAccess class
Implemented the necessary code generation methods for the ArrayAccess class.
The methods implemented are: Emit, EmitAddr, GetMemBytes, and GetMemBytesAddr.
The "Addr" methods were needed to allow the AssignExpr class to properly store
values in an array on assignment. Otherwise, the temporary variables would be
assigned to, and not the array itself.