Remove TODO comments that are no longer valid
Objects can only be allocated on the heap, thus their size on the stack will
always be the size of a single pointer. Thus, there is no need to handle object
sizes on the stack any differently than any other variable.

Polymorphic dynamic dispatching of method calls is already implemented, thus
that comment is out of date.